feat(newsite): add index layout
* this also adds some restructuring
* refactor css
