+++ src/components/layout.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { I18nextProvider } from 'react-i18next';
 import i18n, { getLang } from '../utils/i18n';
 import Seo from '../components/seo';
-import Header from '../components/header';
+import Header from '../components/header/';
 import Footer from '../components/footer';
 
 export default ({ children }) => (
