+++ src/styles/page-intro.scss
@@ -1,0 +1,21 @@
+.page-intro {
+  margin-top: 100px;
+  width: 400px;
+  display: flex;
+  flex-direction: column;
+}
+
+.page-intro__headline {
+  color: #0F2A48;
+  font-weight: 400;
+  > b {
+    font-weight: 600;
+    font-size: 150%;
+    line-height: 150%;
+  }
+}
+
+.page-intro__note {
+  background-color: #F8F9FB;
+  border-radius: 11px;
+}
