+++ src/styles/variables.scss
@@ -1,0 +1,16 @@
+$text-color:                    #999999;
+$brand-primary:                 #3685D6;
+
+$btn-font-weight:               600;
+// Button default
+$btn-default-color:             #CEA000;
+$btn-default-bg:                #FFF2C6;
+$btn-default-border:            transparent;
+
+$btn-primary-color:             #fff;
+$btn-primary-bg:                $brand-primary;
+$btn-primary-border:            transparent;
+
+$btn-info-color:                $text-color;
+$btn-info-bg:                   transparent;
+$btn-info-border:               transparent;
